requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
attrs==24.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.213
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.3
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.1.0
    # via aws-cdk-lib
aws-cdk-cloud-assembly-schema==38.0.1
    # via aws-cdk-lib
aws-cdk-lib==2.171.1
    # via -r requirements.in
cattrs==24.1.2
    # via jsii
constructs==10.4.2
    # via
    #   -r requirements.in
    #   aws-cdk-lib
importlib-resources==6.4.5
    # via jsii
jsii==1.105.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0
    # via jsii
six==1.17.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.12.2
    # via jsii
